Bring back Tool Examples listing, refs #6091.
First, some history: the previous version of the site used a certain content
layout in which you had a Folder with some Tool Examples. This folder also
had a Page and this page was set as the default view for the folder. When
migrating to newer version, all these Pages were transformed into Tool
Examples. This means that the part of the Page template that showed a listing
of Tool Examples needs to be moved to the Tool Example template. This commit
takes care of this.

Apart from that, it adds viewlets.xml to the upgrade step, so you don't need
to manually re-import it in portal_setup.